refactor: move calculate_tax_impact() and its shared models to counties/common
AssessmentHistory, TaxUnitRate, and PropertyJurisdictionExemption all
carry a county column and are documented (wayfinder ticket #9) as a
tax-impact data model reused verbatim across counties -- genuinely
county-neutral. But they, and calculate_tax_impact() that consumes
them, were physically housed inside counties/harris/, so Brazos had a
hard, in-process import dependency on Harris's package that
CountyAdapter itself never declared to callers.

Moves the three models into counties/common/tax_models.py and
calculate_tax_impact()/TaxImpactResult into counties/common/tax_impact.py.
Each model's Meta.app_label stays pinned to "data" (Harris's pinned
app label, see counties/harris/apps.py), so table names and migration
history are untouched -- confirmed via `makemigrations --check`
reporting no changes. Migrations for these models still live in
counties/harris/migrations/, since Django resolves migration history
by app label, not by where the model class's module lives.

Both adapters now import calculate_tax_impact from counties.common
instead of Brazos reaching into Harris's package by name.
counties/common/history.py's AssessmentHistory import is now a real
top-level import instead of the lazy in-function workaround it needed
while the model lived in a county app.

Surfaces from the architecture review (/improve-codebase-architecture,
2026-08-09), issue #15.
