+++ counties/harris/management/commands/import_hcad_jur_exempt.py
@@ -75,7 +75,8 @@ from django.conf import settings
 from django.core.management.base import BaseCommand, CommandError
 from django.db import connection, transaction
 
-from counties.harris.models import PropertyRecord, TaxUnitRate
+from counties.common.tax_models import TaxUnitRate
+from counties.harris.models import PropertyRecord
 
 logger = logging.getLogger(__name__)
