+++ counties/harris/models.py
@@ -169,112 +169,8 @@ class ExtraFeature(models.Model):
         return f"{self.feature_description} for {self.account_number}"
 
 
-# Shared by AssessmentHistory / TaxUnitRate / PropertyJurisdictionExemption: these
-# three tables are the tax-impact data model reused verbatim across counties (see
-# wayfinder ticket #9 on the issue tracker). account_number/prop_id and tax_unit_code
-# formats aren't guaranteed collision-proof across counties, so county disambiguates
-# real key uniqueness rather than relying on incidental format differences.
-COUNTY_CHOICES = [
-    ("harris", "Harris"),
-    ("brazos", "Brazos"),
-]
-
-
-class AssessmentHistory(models.Model):
-    """Year-based assessed value history for real properties."""
-
-    account_number = models.CharField(max_length=20, db_index=True)
-    tax_year = models.IntegerField(db_index=True)
-    county = models.CharField(
-        max_length=16, choices=COUNTY_CHOICES, default="harris", db_index=True
-    )
-    assessed_value = models.DecimalField(max_digits=14, decimal_places=2, null=True, blank=True)
-    appraised_value = models.DecimalField(max_digits=14, decimal_places=2, null=True, blank=True)
-    market_value = models.DecimalField(max_digits=14, decimal_places=2, null=True, blank=True)
-    prior_appraised_value = models.DecimalField(
-        max_digits=14, decimal_places=2, null=True, blank=True
-    )
-    prior_market_value = models.DecimalField(max_digits=14, decimal_places=2, null=True, blank=True)
-    new_construction_value = models.DecimalField(
-        max_digits=14, decimal_places=2, null=True, blank=True
-    )
-    cap_account = models.CharField(max_length=20, blank=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    class Meta:
-        ordering = ["-tax_year"]
-        constraints = [
-            models.UniqueConstraint(
-                fields=["account_number", "tax_year", "county"],
-                name="unique_assessment_history_per_year",
-            )
-        ]
-
-    def __str__(self):
-        return f"{self.account_number} ({self.tax_year})"
-
-
-class TaxUnitRate(models.Model):
-    """Annual tax rate by taxing unit code."""
-
-    tax_year = models.IntegerField(db_index=True)
-    tax_unit_code = models.CharField(max_length=32, db_index=True)
-    county = models.CharField(
-        max_length=16, choices=COUNTY_CHOICES, default="harris", db_index=True
-    )
-    tax_unit_name = models.CharField(max_length=255, blank=True)
-    adopted_rate = models.DecimalField(max_digits=12, decimal_places=8)
-    source = models.CharField(max_length=64, blank=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=["tax_year", "tax_unit_code", "county"],
-                name="unique_tax_rate_per_unit_year",
-            )
-        ]
-        indexes = [
-            models.Index(fields=["tax_year", "tax_unit_code"]),
-        ]
-
-    def __str__(self):
-        return f"{self.tax_unit_code} {self.tax_year}: {self.adopted_rate}"
-
-
-class PropertyJurisdictionExemption(models.Model):
-    """Jurisdiction/exemption row linked to an account and year."""
-
-    account_number = models.CharField(max_length=20, db_index=True)
-    tax_year = models.IntegerField(db_index=True)
-    county = models.CharField(
-        max_length=16, choices=COUNTY_CHOICES, default="harris", db_index=True
-    )
-    tax_unit_code = models.CharField(max_length=32, db_index=True)
-    tax_unit_name = models.CharField(max_length=255, blank=True)
-    exemption_code = models.CharField(max_length=32, blank=True)
-    exemption_description = models.CharField(max_length=255, blank=True)
-    exemption_amount = models.DecimalField(max_digits=14, decimal_places=2, null=True, blank=True)
-    exemption_percent = models.DecimalField(max_digits=8, decimal_places=4, null=True, blank=True)
-    taxable_value = models.DecimalField(max_digits=14, decimal_places=2, null=True, blank=True)
-    assessed_value = models.DecimalField(max_digits=14, decimal_places=2, null=True, blank=True)
-    source = models.CharField(max_length=64, blank=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=["account_number", "tax_year", "tax_unit_code", "exemption_code", "county"],
-                name="unique_jur_exemption_per_unit_code_year",
-            )
-        ]
-        indexes = [
-            models.Index(fields=["account_number", "tax_year"]),
-            models.Index(fields=["tax_year", "tax_unit_code"]),
-        ]
-
-    def __str__(self):
-        return f"{self.account_number} {self.tax_year} {self.tax_unit_code}"
+# AssessmentHistory, TaxUnitRate, and PropertyJurisdictionExemption moved to
+# counties/common/tax_models.py -- they're shared, county-scoped tables, not
+# Harris-owned (wayfinder ticket #9). Their Meta.app_label is still pinned to
+# "data" (this app's label), so nothing about the tables, migration history,
+# or content types changed; only where the Python class definitions live.
